+++ torchtitan/experiments/llama4/model/args.py
@@ -12,8 +12,8 @@ from torch import nn
 from torchtitan.config import JobConfig
 
 from torchtitan.models.moe import MoEArgs
+from torchtitan.models.utils import get_moe_model_nparams_and_flops
 from torchtitan.protocols import BaseModelArgs
-from torchtitan.protocols.model import get_moe_nparams_and_flops
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability
 
@@ -75,4 +75,4 @@ class TransformerModelArgs(BaseModelArgs):
     def get_nparams_and_flops(
         self, model: nn.Module, seq_len: int
     ) -> tuple[int, float]:
-        return get_moe_nparams_and_flops(self, model, seq_len)
+        return get_moe_model_nparams_and_flops(self, model, seq_len)
